feat: move ParseJsonToDict ot helpers
