+++ Assets/PlayroomKit/modules/Helpers/Helpers.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using SimpleJSON;
 using UnityEngine;
 
@@ -213,6 +214,31 @@ namespace Playroom
             profileData.photo = jsonNode["photo"].Value;
 
             return profileData;
+        }
+        
+        private static Dictionary<string, T> ParseJsonToDictionary<T>(string jsonString)
+        {
+            var dictionary = new Dictionary<string, T>();
+            var jsonNode = JSON.Parse(jsonString);
+
+            foreach (var kvp in jsonNode.AsObject)
+            {
+                T value = default; // Initialize the value to default value of T
+
+                // Parse the JSONNode value to the desired type (T)
+                if (typeof(T) == typeof(float))
+                    value = (T)(object)kvp.Value.AsFloat;
+                else if (typeof(T) == typeof(int))
+                    value = (T)(object)kvp.Value.AsInt;
+                else if (typeof(T) == typeof(bool))
+                    value = (T)(object)kvp.Value.AsBool;
+                else
+                    Debug.LogError("Unsupported type: " + typeof(T).FullName);
+
+                dictionary.Add(kvp.Key, value);
+            }
+
+            return dictionary;
         }
